% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{create_spider}
\alias{create_spider}
\title{Create and run a Spider}
\usage{
create_spider(name, start_urls, callback, allowed_domains, custom_settings)
}
\arguments{
\item{name}{The name of the spider [character]}

\item{start_urls}{vector with initial urls [vector of character]}

\item{callback}{the function to handle the response callback from start_urls requests,
must receive one parameter ``response`` that is a data.frame with the response data [function]}

\item{allowed_domains}{vector with allowed domains (optional) [vector of character]}

\item{custom_settings}{custom scrapy settings (optional) [data.frame]}
}
\description{
Check the ScrapyStreaming and Scrapy documentations for more information about these parameters.
}
\examples{
parse <- function(response) {
 print(response$headers)
}

create_spider(name = "sample",
              start_urls = "http://example.com",
              callback = parse,
              allowed_domains = c("example.com", "dmoz.org"),
              custom_settings = data.frame("SOME_SETTING" = "some value"))

}

