% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{handle}
\alias{handle}
\title{Handles the input message}
\usage{
handle(execute)
}
\arguments{
\item{execute}{if TRUE (the default) will execute the message as follows:
* ready: if the status is ready, continue the
* response: call the callback with the message as the first parameter
* error: stop the execution of the process and show the spider error
* exception: stop the execution of the process and show the scrapy exception}
}
\description{
Handles the input message
}

