% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{parse_input}
\alias{parse_input}
\title{Read and returns a stdin line}
\usage{
parse_input(raw, ...)
}
\arguments{
\item{raw}{if TRUE, returns the raw input line. If FALSE (the default), parse it using jsonlite::fromJSON}

\item{...}{extra arguments to jsonlite::fromJSON}
}
\value{
the json message sent to stdout
}
\description{
Read and returns a stdin line
}

