% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{send_from_response_request}
\alias{send_from_response_request}
\title{Create an open a new request}
\usage{
send_from_response_request(url, callback, from_response_request, base64, method,
  meta, body, headers, cookies, encoding, priority, dont_filter)
}
\arguments{
\item{url}{request url [character]}

\item{callback}{the function to handle the response callback,
must receive one parameter ``response`` that is a data.frame with the response data [function]}

\item{from_response_request}{data to the new request. May contain fields to the request and to the form. [data.frame]}

\item{base64}{bool, if True, the response body will be encoded with base64 (optional) [logical]}

\item{method}{request method (optional) [character]}

\item{meta}{metadata to the request (optional) [data.frame]}

\item{body}{the request body (optional) [character]}

\item{headers}{request headers (optional) [data.frame]}

\item{cookies}{request cookies  (optional) [data.frame]}

\item{encoding}{request encoding  (optional) [character]}

\item{priority}{request priority  (optional) [numeric]}

\item{dont_filter}{indicates that this request should not be filtered by the scheduler (optional) [logical]}
}
\description{
Create an open a new request and uses its response to open a new request using scrapy's FormRequest.from_response and ``form_request`` data.
Check the ScrapyStreaming and Scrapy documentations for more information about these parameters.
}
\examples{
my_callback <- function (response) {
 print(response)
}
send_from_response_request("http://example.com", my_callback,
             data.frame(formcss = "#login_form", formdata = data.frame(user = "admin", pass = "1")))

}

