% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{send_log}
\alias{send_log}
\title{Send a log to Scrapy Streaming}
\usage{
send_log(message, level)
}
\arguments{
\item{message}{the log message}

\item{level}{log level. Must be one of 'CRITICAL', 'ERROR', 'WARNING', 'DEBUG', and 'INFO', defaults to DEBUG}
}
\description{
Send a log to Scrapy Streaming
}
\examples{
send_log("log works!")
send_log("logging error :(", "error")

}

