% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{send_request}
\alias{send_request}
\title{Create an open a new request}
\usage{
send_request(url, callback, base64, method, meta, body, headers, cookies,
  encoding, priority, dont_filter)
}
\arguments{
\item{url}{request url [character]}

\item{callback}{the function to handle the response callback,
must receive one parameter ``response`` that is a data.frame with the response data [function]}

\item{base64}{bool, if True, the response body will be encoded with base64 (optional) [logical]}

\item{method}{request method (optional) [character]}

\item{meta}{metadata to the request (optional) [data.frame]}

\item{body}{the request body (optional) [character]}

\item{headers}{request headers (optional) [data.frame]}

\item{cookies}{request cookies  (optional) [data.frame]}

\item{encoding}{request encoding  (optional) [character]}

\item{priority}{request priority  (optional) [numeric]}

\item{dont_filter}{indicates that this request should not be filtered by the scheduler (optional) [logical]}
}
\description{
Check the ScrapyStreaming and Scrapy documentations for more information about these parameters.
}
\examples{
my_callback <- function (response) {
 print(response)
}
send_request("http://example.com", my_callback)

}

